+++ api/ai_services.py
@@ -6,7 +6,7 @@ from bson import ObjectId
 
 from pydantic import BaseModel, Field
 
-from ai_config.ai_constants import (
+from api.constants.ai_constants import (
     EMBEDDING_CHUNK_SIZE,
     EMBEDDING_OVERLAP,
 )
@@ -168,63 +168,3 @@ class GradeConfidenceLevel(BaseModel):
         ge=0.0,
         le=1.0,
     )
-
-# TODO: DEPRECATED
-''''
-# created message class to keep track of messages that build up a Conversation
-class Message:
-    def __init__(self, role, content, timestamp=None):
-        self.role = role
-        self.content = content
-        self.timestamp = timestamp or datetime.now().isoformat()
-    def add_message(self, role, content):
-        message = Message(role, content)
-        self.messages.append(message)
-
-        # translation layer
-        if role == "assistant":
-            malay_translation = translate_en_to_ms(content)
-            chinese_translation = content
-
-            self.translations.append(
-                {
-                    "message_id": len(self.messages) - 1,
-                    "translations": [
-                        {"language": "en", "text": content},
-                        {
-                            "language": "ms_MY",
-                            "text": malay_translation.get("text", ""),
-                        },
-                        {
-                            "language": "zh_CN",
-                            "text": chinese_translation.get("text", ""),
-                        },
-                        {
-                            "language": "zh_TW",
-                            "text": chinese_translation.get("text", ""),
-                        },
-                    ],
-                }
-            )
-            return message
-    def get_conversation_history(self):
-        return [
-            {"role": msg.role, "content": msg.content, "timestamp": msg.timestamp}
-            for msg in self.messages
-        ]
-    # save the output to a dict for using API
-    def to_dict(self):
-        return {
-            "_id": self._id,
-            "session_id": self.session_id,
-            "admin_id": self.admin_id,
-            "user_id": self.user_id,
-            "bot_id": self.bot_id,
-            "timestamp": self.timestamp,
-            "messages": [
-                {"role": msg.role, "content": msg.content, "timestamp": msg.timestamp}
-                for msg in self.messages
-            ],
-            "translations": self.translations,
-        }
-'''
