--- api/feedback.py
@@ -1,5302 +1,0 @@
-[
-   {
-        "id": "0201",
-        "question": {
-            "text": "I have a question related to account access. Can you help clarify it? | I forgot to fill in my Joker ID during the transfer",
-            "variations": [
-                "Can you explain how account access works?",
-                "I need assistance regarding account access. What should I do?",
-                "My transaction related to account access isn’t going through. Can you check?",
-                "I am facing an issue with account access. How can I resolve it?"
-            ],
-            "intent": "account_access",
-            "languages": {
-                "en": "I have a question related to account access. Can you help clarify it? | I forgot to fill in my Joker ID during the transfer",
-                "ms": "\"Saya lupa isi ID Joker waktu transfer\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss ~\\r\\nPlease fill in your JOKER ID during online transfers in the future ~\\r\\n❗️Fill in the designated place❗️\\r\\n✅ Recipient Reference\\r\\n✅ Other Payment Details\\r\\n▶️ 6540JKM ◀️\\r\\nThank you, boss ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Hi Bosskuu ~\\r\\nMinta Masa Depan Isi ID JOKER Anda Semasa Buat Online Transfer Ya ~\\r\\n❗️ISI DI TEMPAT ❗️\\r\\n✅ Recipient Reference\\r\\n✅Other Payment Details\\r\\n▶️ 6540JKM ◀️\\r\\nTerima Kasih Boss ❤️\"",
-                "cn": "\"嗨，老闆~\\r\\n未來請在線上轉帳時填寫您的JOKER ID哦 ~\\r\\n❗️填寫在指定位置❗️\\r\\n✅ 收款人參考\\r\\n✅ 其他支付細節\\r\\n▶️ 6540JKM ◀️\\r\\n謝謝老闆 ❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "account_management"
-            ],
-            "subCategory": "account_access",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818488Z",
-            "lastUpdated": "2025-02-27T06:39:18.818490Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0202",
-        "question": {
-            "text": "I have a question related to registration. Can you help clarify it? | Boss, if I change my phone number, can I still use my original Joker account?",
-            "variations": [
-                "Can you explain how registration works?",
-                "I need assistance regarding registration. What should I do?",
-                "My transaction related to registration isn’t going through. Can you check?",
-                "I am facing an issue with registration. How can I resolve it?"
-            ],
-            "intent": "registration",
-            "languages": {
-                "en": "I have a question related to registration. Can you help clarify it? | Boss, if I change my phone number, can I still use my original Joker account?",
-                "ms": "\"bos klu tukar no hp joker empunya sy blh guna jg ka\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Change it here first, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"tukar dkt sini dulu bos\"",
-                "cn": "\"先在這裡換哦，老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "account_management"
-            ],
-            "subCategory": "registration",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818501Z",
-            "lastUpdated": "2025-02-27T06:39:18.818504Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0203",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | How is it calculated?",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | How is it calculated?",
-                "ms": "\"mcm mana kira\"",
-                "cn": "\"怎麼計算？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, only welcome message) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅歡迎訊息)",
-                "cn": "(無明確解決方案，僅歡迎訊息)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818516Z",
-            "lastUpdated": "2025-02-27T06:39:18.818520Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0204",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | The withdrawal money hasn’t come in yet",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | The withdrawal money hasn’t come in yet",
-                "ms": "\"Duit withdraw belum masuk lagi\"",
-                "cn": "\"提款的錢還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Wait 5 minutes, if it doesn’t come in, contact us again, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"tunggu 5min tak masuk roger balik ye bos\"",
-                "cn": "\"等5分鐘如果沒進來請再次聯繫哦，老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818556Z",
-            "lastUpdated": "2025-02-27T06:39:18.818558Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0205",
-        "question": {
-            "text": "I have a question related to game access. Can you help clarify it? | Is there a problem with Mega?",
-            "variations": [
-                "Can you explain how game access works?",
-                "I need assistance regarding game access. What should I do?",
-                "My transaction related to game access isn’t going through. Can you check?",
-                "I am facing an issue with game access. How can I resolve it?"
-            ],
-            "intent": "game_access",
-            "languages": {
-                "en": "I have a question related to game access. Can you help clarify it? | Is there a problem with Mega?",
-                "ms": "\"mega ada masalah ke\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please send a screenshot, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"screenshot mari bso\"",
-                "cn": "\"請發截圖過來，老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "game_access",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818567Z",
-            "lastUpdated": "2025-02-27T06:39:18.818569Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0206",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | What’s going on, it hasn’t come in yet",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | What’s going on, it hasn’t come in yet",
-                "ms": "\"mcm mn ni .. xde masuk pun\"",
-                "cn": "\"怎麼回事，還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, only welcome message) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅歡迎訊息)",
-                "cn": "(無明確解決方案，僅歡迎訊息)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818580Z",
-            "lastUpdated": "2025-02-27T06:39:18.818582Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0207",
-        "question": {
-            "text": "I have a question related to transaction records. Can you help clarify it? | RM 102.00 hasn’t come in yet, how many days has it been?",
-            "variations": [
-                "Can you explain how transaction records works?",
-                "I need assistance regarding transaction records. What should I do?",
-                "My transaction related to transaction records isn’t going through. Can you check?",
-                "I am facing an issue with transaction records. How can I resolve it?"
-            ],
-            "intent": "transaction_records",
-            "languages": {
-                "en": "I have a question related to transaction records. Can you help clarify it? | RM 102.00 hasn’t come in yet, how many days has it been?",
-                "ms": "\"rm 102.00 belum masuk...dah brp hari ni ..\"",
-                "cn": "\"102令吉還沒進來，已經幾天了？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please check your bank, boss, thank you, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"minta bos check bank anda ya tq boosku\"",
-                "cn": "\"請老闆檢查您的銀行哦，謝謝老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "transaction_records",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818590Z",
-            "lastUpdated": "2025-02-27T06:39:18.818596Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0208",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Why can’t I press it, I’ve already entered it?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Why can’t I press it, I’ve already entered it?",
-                "ms": "\"knp tk blh tekan saya dh masukkan dh\"",
-                "cn": "\"為什麼不能按，我已經輸入進去了？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Just use fast payment If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"pakai fastpayment ja\"",
-                "cn": "\"只用快付就好\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818613Z",
-            "lastUpdated": "2025-02-27T06:39:18.818616Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0209",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred to your account, not using IBG",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred to your account, not using IBG",
-                "ms": "\"aku tranfer acc ko bukan pakai ibg\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please do one fast payment top-up first\\r\\nContact here when done If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sila buat satu topup fast payment dulu\\r\\nsiap roger sini\"",
-                "cn": "\"請先進行一次快付充值\\r\\n完成後在這裡聯繫\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818623Z",
-            "lastUpdated": "2025-02-27T06:39:18.818625Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0210",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I don’t understand how to top up with fast payment",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I don’t understand how to top up with fast payment",
-                "ms": "\"Saya tak faham cara topup fast payment\"",
-                "cn": "\"我不懂怎麼用快付充值\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "On the TOPUP page\\r\\nPress fast payment\\r\\nThere are TNG, Maybank, and other options\\r\\nLog in, pay\\r\\nThe money will come in automatically If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"dekat HALAMAN TOPUP\\r\\ntekan fast payment\\r\\nada tng , maybank semua tu\\r\\nlogin masuk , bayar\\r\\nduit auto masuk\"",
-                "cn": "\"在充值頁面\\r\\n按快付\\r\\n有TNG、Maybank等選項\\r\\n登錄進入，付款\\r\\n錢會自動進來\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818632Z",
-            "lastUpdated": "2025-02-27T06:39:18.818635Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0211",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | What about my RM 100 now?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | What about my RM 100 now?",
-                "ms": "\"skrg duit aku mcm mn 100 tu\"",
-                "cn": "\"現在我的100令吉怎麼樣了？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please do one fast payment top-up first\\r\\nContact here when done If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sila buat satu topup fast payment dulu\\r\\nsiap roger sini\"",
-                "cn": "\"請先進行一次快付充值\\r\\n完成後在這裡聯繫\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818646Z",
-            "lastUpdated": "2025-02-27T06:39:18.818649Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0212",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Good afternoon, boss, I want to ask how to get back the RM 5 I topped up earlier by mistake",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Good afternoon, boss, I want to ask how to get back the RM 5 I topped up earlier by mistake",
-                "ms": "\"Selamat petang bos mahu tanya mcm mana nak dapatkan rm5 yg sy topup tdi ya sy silap\"",
-                "cn": "\"下午好，老闆，想問怎麼拿回我剛才錯充的5令吉？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Send both receipts here first\\r\\nDon’t upload If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"dua2 resit send sini dulu\\r\\njagan uplaod\"",
-                "cn": "\"先把兩張收據發到這裡\\r\\n不要上傳\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818659Z",
-            "lastUpdated": "2025-02-27T06:39:18.818661Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0213",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | How do I get back the RM 5 I topped up earlier by mistake? The second time I topped up RM 10, it went through, but the RM 5 didn’t come in",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | How do I get back the RM 5 I topped up earlier by mistake? The second time I topped up RM 10, it went through, but the RM 5 didn’t come in",
-                "ms": "\"mcm mana nak dapatkan rm5 yg sy topup tdi ya sy silap .lepas yg ke 2 kali sy topup lagi rm 10 ni dapat yg rm 5 tu tak dapat masok\"",
-                "cn": "\"怎麼拿回我剛才錯充的5令吉，第二次我又充了10令吉，這次進了，但5令吉沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Screenshot our chat, send it to the top-up section, and fill in RM 15 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"screenshot chat kita, send ke topup dan isi rm15\"",
-                "cn": "\"截圖我們的對話，發到充值頁面並填入15令吉\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818672Z",
-            "lastUpdated": "2025-02-27T06:39:18.818675Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0214",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | This hasn’t come in yet, boss?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | This hasn’t come in yet, boss?",
-                "ms": "\"ne x masuk ka bos\"",
-                "cn": "\"這還沒進來嗎，老闆？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user confirmed it’s in) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶確認已進入)",
-                "cn": "(無明確解決方案，僅用戶確認已進入)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818682Z",
-            "lastUpdated": "2025-02-27T06:39:18.818684Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0215",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred wrongly",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred wrongly",
-                "ms": "\"Sy slh tranfer\"",
-                "cn": "\"我轉錯了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user provided info and confirmed it’s in) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶提供資訊後確認已進入)",
-                "cn": "(無明確解決方案，僅用戶提供資訊後確認已進入)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818701Z",
-            "lastUpdated": "2025-02-27T06:39:18.818704Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0216",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | But the money hasn’t come in yet",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | But the money hasn’t come in yet",
-                "ms": "\"tapi duit belom masuk\"",
-                "cn": "\"但錢還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please be patient, boss, we’ll check later If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sabar ya bos semak nanti ya\"",
-                "cn": "\"請耐心等待，稍後會檢查哦，老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818712Z",
-            "lastUpdated": "2025-02-27T06:39:18.818714Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0217",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Top-up hasn’t come in",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Top-up hasn’t come in",
-                "ms": "\"Topup xmasuk\"",
-                "cn": "\"充值沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user confirmed it’s in) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶確認已進入)",
-                "cn": "(無明確解決方案，僅用戶確認已進入)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818731Z",
-            "lastUpdated": "2025-02-27T06:39:18.818733Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0218",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Deposit hasn’t come in",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Deposit hasn’t come in",
-                "ms": "\"deposit belum masuk\"",
-                "cn": "\"存款還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, only welcome message) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅歡迎訊息)",
-                "cn": "(無明確解決方案，僅歡迎訊息)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818745Z",
-            "lastUpdated": "2025-02-27T06:39:18.818747Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0219",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Wrote the wrong reference number",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Wrote the wrong reference number",
-                "ms": "\"salah tulis no reference\"",
-                "cn": "\"寫錯了參考號\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please upload your bank slip via (WALLET -> TOPUP), the process takes 3-5 minutes, thank you, boss, awesome! ❤️❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sila uplaod slip bankin anda melalui (WALLET -> TOPUP ) proses ambil masa dalam 3-5min ,tq boss, mantap❤️❤️\"",
-                "cn": "\"請通過（WALLET -> TOPUP）上傳您的銀行單據，處理時間為3-5分鐘，謝謝老闆，太棒了❤️❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818754Z",
-            "lastUpdated": "2025-02-27T06:39:18.818756Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0220",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, once again the fast payment didn’t come in",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, once again the fast payment didn’t come in",
-                "ms": "\"lagi sekali bos tak masuk fast payment\"",
-                "cn": "\"老闆，又一次快付沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user confirmed it’s in) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶確認已進入)",
-                "cn": "(無明確解決方案，僅用戶確認已進入)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818763Z",
-            "lastUpdated": "2025-02-27T06:39:18.818765Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0221",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Still pending, stupid",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Still pending, stupid",
-                "ms": "\"pending lagi bodoh\"",
-                "cn": "\"還在等待，真蠢\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss, if you’ve already uploaded the receipt to top-up pending, please be patient, it’s being processed, thank you, boss~ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"hi boss kalau suda upload resit ke topup pending, minta sabar ya otw proses suda. tq boskuu~\"",
-                "cn": "\"嗨，老闆，如果已經上傳收據到充值待處理，請耐心等待，正在處理中，謝謝老闆~\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818777Z",
-            "lastUpdated": "2025-02-27T06:39:18.818780Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0222",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | Hello, boss, tonight I won RM 76.25, but the total is RM 57.34",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | Hello, boss, tonight I won RM 76.25, but the total is RM 57.34",
-                "ms": "\"Hello boss mlm ni menang rm76. 25 tapi jumlahnya rm57. 34.\"",
-                "cn": "\"老闆，今晚我贏了76.25令吉，但總額是57.34令吉\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please check your transaction history again If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sila semak balik transaction history anda ya\"",
-                "cn": "\"請再次檢查您的交易記錄哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818792Z",
-            "lastUpdated": "2025-02-27T06:39:18.818794Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0223",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Sorry, I wrote RM 50.04 wrongly, it came in as RM 55.04",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Sorry, I wrote RM 50.04 wrongly, it came in as RM 55.04",
-                "ms": "\"sorry salah tulis rm50.04 masuk rm55.04\"",
-                "cn": "\"抱歉寫錯了50.04令吉，進來的是55.04令吉\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "The amount earlier was written wrongly, please upload again and write the amount as RM 55.04 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Amount tadi silap tulis , sila upload semula dan amount tulis RM 55.04\"",
-                "cn": "\"剛才金額寫錯了，請重新上傳並將金額寫為55.04令吉\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818803Z",
-            "lastUpdated": "2025-02-27T06:39:18.818805Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0224",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | I withdrew RM 50.08, it hasn’t come in yet",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | I withdrew RM 50.08, it hasn’t come in yet",
-                "ms": "\"Sy ada withdraw rm50. 08 masuk lg\"",
-                "cn": "\"我提了50.08令吉，還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Withdrawal of winnings takes 5-30 minutes, boss, please wait a moment ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Pengeluaran wang menang ambil masa dalam 5-30 min yee boss ,sila tunggu sebentar yeee❤️\"",
-                "cn": "\"贏錢的提款需要5-30分鐘哦，老闆，請稍等一下哦❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818888Z",
-            "lastUpdated": "2025-02-27T06:39:18.818893Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0225",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred to CIMB but it hasn’t come in",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred to CIMB but it hasn’t come in",
-                "ms": "\"Saya bankin ke CIMB tapi tak masuk\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss, if you transferred to [CIMB], please select [CIMB] and then upload your receipt, thank you for your cooperation, boss 😘 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Hi boss , jika boss bankin ke [ CIMB ] sila pilih [ CIMB ] baru upload resit anda ye , Terima Kasih Atas Kerjasama ya bosss😘\"",
-                "cn": "\"嗨，老闆，如果老闆轉帳到[CIMB]，請選擇[CIMB]再上傳您的收據，謝謝合作哦，老闆😘\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818911Z",
-            "lastUpdated": "2025-02-27T06:39:18.818915Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0226",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Not in yet",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Not in yet",
-                "ms": "\"BLM msuk\"",
-                "cn": "\"還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please send the receipt here If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"resit send sini\"",
-                "cn": "\"請把收據發到這裡\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818925Z",
-            "lastUpdated": "2025-02-27T06:39:18.818928Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0227",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I haven’t received it yet",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I haven’t received it yet",
-                "ms": "\"还没有收到\"",
-                "cn": "\"還沒收到\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Wait 10 minutes, if it doesn’t come in, contact us If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"tunggu 10min xmasuk roger ya\"",
-                "cn": "\"等10分鐘如果沒進來請聯繫哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818942Z",
-            "lastUpdated": "2025-02-27T06:39:18.818945Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0228",
-        "question": {
-            "text": "I have a question related to game information. Can you help clarify it? | What are some tips for Mega?",
-            "variations": [
-                "Can you explain how game information works?",
-                "I need assistance regarding game information. What should I do?",
-                "My transaction related to game information isn’t going through. Can you check?",
-                "I am facing an issue with game information. How can I resolve it?"
-            ],
-            "intent": "game_information",
-            "languages": {
-                "en": "I have a question related to game information. Can you help clarify it? | What are some tips for Mega?",
-                "ms": "\"apa tips mega?\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, only user dissatisfaction) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶不滿)",
-                "cn": "(無明確解決方案，僅用戶不滿)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "game"
-            ],
-            "subCategory": "game_information",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818957Z",
-            "lastUpdated": "2025-02-27T06:39:18.818959Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0229",
-        "question": {
-            "text": "I have a question related to game information. Can you help clarify it? | Is there an MKT (market) today?",
-            "variations": [
-                "Can you explain how game information works?",
-                "I need assistance regarding game information. What should I do?",
-                "My transaction related to game information isn’t going through. Can you check?",
-                "I am facing an issue with game information. How can I resolve it?"
-            ],
-            "intent": "game_information",
-            "languages": {
-                "en": "I have a question related to game information. Can you help clarify it? | Is there an MKT (market) today?",
-                "ms": "\"hari ada MKT ke?\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user confirmed) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶確認)",
-                "cn": "(無明確解決方案，僅用戶確認)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "game"
-            ],
-            "subCategory": "game_information",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818967Z",
-            "lastUpdated": "2025-02-27T06:39:18.818970Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0230",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, the amount is wrong, it’s 150, not 100",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, the amount is wrong, it’s 150, not 100",
-                "ms": "\"silap jumlah bos...150..bukan 100\"",
-                "cn": "\"老闆，金額錯了，是150，不是100\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "The amount earlier was written wrongly, please upload again and write the amount as RM 150 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Amount tadi silap tulis , sila upload semula dan amount tulis RM 150\"",
-                "cn": "\"剛才金額寫錯了，請重新上傳並將金額寫為150令吉\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818977Z",
-            "lastUpdated": "2025-02-27T06:39:18.818979Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0231",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Sent the wrong receipt",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Sent the wrong receipt",
-                "ms": "\"Silap send resit\"",
-                "cn": "\"發錯收據了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user confirmed it’s resolved) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶確認已解決)",
-                "cn": "(無明確解決方案，僅用戶確認已解決)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.818991Z",
-            "lastUpdated": "2025-02-27T06:39:18.818993Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0232",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Hello, boss, there was an error in the last submission ❌ Previously filled wrongly: RM 710 ✅ Please fill: RM 170",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Hello, boss, there was an error in the last submission ❌ Previously filled wrongly: RM 710 ✅ Please fill: RM 170",
-                "ms": "\"Hello Bossku ~ Submit Tadi Ada Kesilapan ❌ Tadi Silap Isi : RM710 ✅ Sila Isi : RM170\"",
-                "cn": "\"嗨，老闆，剛才提交有誤 ❌ 剛才錯填：710令吉 ✅ 請填：170令吉\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hello, boss, there was an error in the last submission ❌ Previously filled wrongly: RM 710 ✅ Please fill: RM 170 Resubmit, thank you, boss ~ ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Hello Bossku ~ Submit Tadi Ada Kesilapan ❌ Tadi Silap Isi : RM710 ✅ Sila Isi : RM170 Submit Semula ye ~ Thank You Bosskuu ~ ❤️\"",
-                "cn": "\"嗨，老闆，剛才提交有誤 ❌ 剛才錯填：710令吉 ✅ 請填：170令吉 請重新提交哦 ~ 謝謝老闆 ~ ❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.819009Z",
-            "lastUpdated": "2025-02-27T06:39:18.819011Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0233",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Sorry, boss, I’ve already transferred here",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Sorry, boss, I’ve already transferred here",
-                "ms": "\"sorry bos....dh transfer kat sini\"",
-                "cn": "\"抱歉，老闆，已經轉帳到這裡了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, only welcome message) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅歡迎訊息)",
-                "cn": "(無明確解決方案，僅歡迎訊息)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.819018Z",
-            "lastUpdated": "2025-02-27T06:39:18.819020Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0234",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Hi, boss, this isn’t our company’s account, did you upload the wrong receipt?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Hi, boss, this isn’t our company’s account, did you upload the wrong receipt?",
-                "ms": "\"hi boss ini bukan acc company kami ya , adakah boss silap upload resit ya ?\"",
-                "cn": "\"嗨，老闆，這不是我們公司的帳戶哦，老闆是不是上傳錯收據了？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please check again If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"semak balik ya\"",
-                "cn": "\"請再檢查一下哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.819031Z",
-            "lastUpdated": "2025-02-27T06:39:18.819034Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0235",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred but to the wrong account",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred but to the wrong account",
-                "ms": "\"Saya transfer tapi salah akaun\"",
-                "cn": "\"我轉帳但轉錯帳戶了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(SUZAINI BIN GASAN) You can only transfer using this name If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"( SUZAINI BIN GASAN ) hanya dpt guna NAMA atas ni trasnfer saja ya\"",
-                "cn": "\"(SUZAINI BIN GASAN) 只能使用這個名字轉帳哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.819913Z",
-            "lastUpdated": "2025-02-27T06:39:18.819927Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0236",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | Did 1057 win?",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | Did 1057 win?",
-                "ms": "\"1057 kena kan?\"",
-                "cn": "\"1057中了嗎？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss, if the 4D number wins, check the report later at 9 PM, the winning credit will automatically enter your Joker wallet, thank you, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"hi bosskuu kalua no 4d ada kena , semak report nnti 9pm ya kredit menang akan auto masuk wallet joker ada ya tq bosku\"",
-                "cn": "\"嗨，老闆，如果4D號碼中獎，稍後晚上9點檢查報告哦，贏的信用會自動進入Joker錢包，謝謝老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.819955Z",
-            "lastUpdated": "2025-02-27T06:39:18.819959Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0237",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | Boss, what time will I receive the payment?",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | Boss, what time will I receive the payment?",
-                "ms": "\"boss jam berapa dpt pembayaran\"",
-                "cn": "\"老闆，幾點可以收到付款？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please be patient and wait for the report later If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sabar ya tunggu report nanti ya\"",
-                "cn": "\"請耐心等待稍後的報告哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.819972Z",
-            "lastUpdated": "2025-02-27T06:39:18.819975Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0238",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Hi, boss, I withdrew RM 50 earlier, but it hasn’t come in yet",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Hi, boss, I withdrew RM 50 earlier, but it hasn’t come in yet",
-                "ms": "\"hai bos sy ada withdraw rm 50 tadi tpi Blum masuklagi\"",
-                "cn": "\"嗨，老闆，我剛才提了50令吉，但還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user confirmed) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶確認)",
-                "cn": "(無明確解決方案，僅用戶確認)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.819996Z",
-            "lastUpdated": "2025-02-27T06:39:18.820000Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0239",
-        "question": {
-            "text": "I have a question related to game access. Can you help clarify it? | Hi, boss, can you send the Mega888 link?",
-            "variations": [
-                "Can you explain how game access works?",
-                "I need assistance regarding game access. What should I do?",
-                "My transaction related to game access isn’t going through. Can you check?",
-                "I am facing an issue with game access. How can I resolve it?"
-            ],
-            "intent": "game_access",
-            "languages": {
-                "en": "I have a question related to game access. Can you help clarify it? | Hi, boss, can you send the Mega888 link?",
-                "ms": "\"hi boss boleh send link mega888 ?\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss, if you want to play Mega888, follow these steps:\\r\\n\\r\\n📌 How to Play Mega888:\\r\\n1️⃣ Download Mega888 from these links:\\r\\n🔗 https://m.cfbz888.com\\r\\n🔗 http://m.mega266.com\\r\\n🔗 https://m2.mega166.com\\r\\n\\r\\n2️⃣ Open the Joker App & copy the ID & Password for Mega888.\\r\\n3️⃣ Log in to Mega888 using that ID & Password.\\r\\n4️⃣ Go to \"Wallet\" in the Joker App & transfer credit to Mega888.\\r\\n5️⃣ Press \"Start & Launch\" (Android) or open the game manually (iOS).\\r\\n6️⃣ Start playing & enjoy the game! 🎰🔥\\r\\n\\r\\n📌 If you have issues with downloading, logging in, or transferring credit, contact us in live chat, boss! 😊🙏 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Hi boss, kalau nak main Mega888, boleh ikut langkah ini:\\r\\n\\r\\n📌 Cara Main Mega888:\\r\\n1️⃣ Download Mega888 dari link berikut:\\r\\n🔗 https://m.cfbz888.com\\r\\n🔗 http://m.mega266.com\\r\\n🔗 https://m2.mega166.com\\r\\n\\r\\n2️⃣ Buka Joker App & salin ID & Password untuk Mega888.\\r\\n3️⃣ Login ke Mega888 guna ID & Password tersebut.\\r\\n4️⃣ Pergi ke \"Wallet\" dalam Joker App & transfer kredit ke Mega888.\\r\\n5️⃣ Tekan \"Start & Launch\" (Android) atau buka game secara manual (iOS).\\r\\n6️⃣ Mula main & enjoy game! 🎰🔥\\r\\n\\r\\n📌 Kalau ada masalah download, login, atau transfer kredit, boleh roger kami di live chat ya boss! 😊🙏\"",
-                "cn": "\"嗨，老闆，如果想玩Mega888，可以按照以下步驟：\\r\\n\\r\\n📌 如何玩Mega888：\\r\\n1️⃣ 從以下鏈接下載Mega888：\\r\\n🔗 https://m.cfbz888.com\\r\\n🔗 http://m.mega266.com\\r\\n🔗 https://m2.mega166.com\\r\\n\\r\\n2️⃣ 打開Joker應用並複製Mega888的ID和密碼。\\r\\n3️⃣ 使用該ID和密碼登錄Mega888。\\r\\n4️⃣ 在Joker應用中前往“錢包”並將信用轉移到Mega888。\\r\\n5️⃣ 按“開始並啟動”（Android）或手動打開遊戲（iOS）。\\r\\n6️⃣ 開始玩並享受遊戲！🎰🔥\\r\\n\\r\\n📌 如果下載、登錄或轉信用有問題，可以在即時聊天中聯繫我們哦，老闆！😊🙏\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "game_access",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820016Z",
-            "lastUpdated": "2025-02-27T06:39:18.820020Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0240",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I wrote the wrong amount",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I wrote the wrong amount",
-                "ms": "\"Saya tulis amount salah\"",
-                "cn": "\"我寫錯金額了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "The amount earlier was written wrongly, please upload again and write the amount as RM 100 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Amount tadi silap tulis , sila upload semula dan amount tulis RM 100\"",
-                "cn": "\"剛才金額寫錯了，請重新上傳並將金額寫為100令吉\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820038Z",
-            "lastUpdated": "2025-02-27T06:39:18.820041Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0241",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Hi, boss, I made a withdrawal, but it hasn’t come in yet 🤗",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Hi, boss, I made a withdrawal, but it hasn’t come in yet 🤗",
-                "ms": "\"hi bos sya ada withdraw tapi belum masuk lagi 🤗\"",
-                "cn": "\"嗨，老闆，我有提款但還沒進來 🤗\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Withdrawal of winnings takes 5-30 minutes, boss, please wait a moment ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Pengeluaran wang menang ambil masa dalam 5-30 min yee boss ,sila tunggu sebentar yeee❤️\"",
-                "cn": "\"贏錢的提款需要5-30分鐘哦，老闆，請稍等一下哦❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820051Z",
-            "lastUpdated": "2025-02-27T06:39:18.820055Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0242",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | I’ve waited a long time, and the money still hasn’t come in",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | I’ve waited a long time, and the money still hasn’t come in",
-                "ms": "\"Lama sudah saya tumggu belum juga masuk duit\"",
-                "cn": "\"我等了很久，錢還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss, if your withdrawal has been approved, it means it’s being processed. Many people are withdrawing now, so when it’s your turn, it will go into your bank account, don’t worry, please be patient, thank you, boss ^^ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"hi bossku. withdraw anda kalau da approved, maksud dlm process da. skrg ramai cuci jadi nnti sampai giliran boss akan masuk bank anda ya jgn risau . minta boss sabar2 ye tq boss ^^\"",
-                "cn": "\"嗨，老闆，您的提款如果已通過審批，就表示正在處理中。現在很多人提現，所以等到您的順序時會進入您的銀行帳戶哦，別擔心，請老闆耐心等待，謝謝老闆 ^^\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820072Z",
-            "lastUpdated": "2025-02-27T06:39:18.820076Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0243",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Hello, boss, please upload a receipt picture that shows 📌 Successful ✅ 📌 Amount 📌 Time/Date 📌 Recipient",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Hello, boss, please upload a receipt picture that shows 📌 Successful ✅ 📌 Amount 📌 Time/Date 📌 Recipient",
-                "ms": "\"Hello Boss, sila upload GAMBAR RESIT yang boleh nampak 📌 Successful ✅ 📌 JUMLAH 📌 MASA / TARIKH 📌 PENERIMA\"",
-                "cn": "\"老闆，請上傳能看到的收據照片 📌 成功 ✅ 📌 金額 📌 時間/日期 📌 收款人\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hello, boss, please upload a receipt picture that shows\\r\\n\\r\\n📌 Successful ✅\\r\\n📌 Amount\\r\\n📌 Time/Date\\r\\n📌 Recipient\\r\\n\\r\\nResubmit another one, thank you, boss ~ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Hello Boss, sila upload GAMBAR RESIT yang boleh nampak\\r\\n\\r\\n📌 Successful ✅\\r\\n📌 JUMLAH\\r\\n📌 MASA / TARIKH\\r\\n📌 PENERIMA\\r\\n\\r\\nResubmit lain ye, Thank you bossku ~\"",
-                "cn": "\"老闆，請上傳能看到的收據照片\\r\\n\\r\\n📌 成功 ✅\\r\\n📌 金額\\r\\n📌 時間/日期\\r\\n📌 收款人\\r\\n\\r\\n請重新提交另一張哦，謝謝老闆 ~\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820095Z",
-            "lastUpdated": "2025-02-27T06:39:18.820099Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0244",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | Did I hit ibox?",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | Did I hit ibox?",
-                "ms": "\"ada Kene ibox\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss, if the 4D number wins, check the report later at 9 PM, the winning credit will automatically enter your Joker wallet, thank you, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"hi bosskuu kalua no 4d ada kena , semak report nnti 9pm ya kredit menang akan auto masuk wallet joker ada ya tq bosku\"",
-                "cn": "\"嗨，老闆，如果4D號碼中獎，稍後晚上9點檢查報告哦，贏的信用會自動進入Joker錢包，謝謝老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820114Z",
-            "lastUpdated": "2025-02-27T06:39:18.820118Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0245",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Why can’t I withdraw?",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Why can’t I withdraw?",
-                "ms": "\"kenapa saya tidak dpat withdraw\"",
-                "cn": "\"為什麼我不能提款？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "The password is wrong\\r\\nThe password is the 6-digit JOKER login password If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"password salah ya\\r\\npassword tu 6 digit login JOKER ya\"",
-                "cn": "\"密碼錯了哦\\r\\n密碼是6位數的JOKER登錄密碼哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820159Z",
-            "lastUpdated": "2025-02-27T06:39:18.820165Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0246",
-        "question": {
-            "text": "I have a question related to game access. Can you help clarify it? | Boss, Mega can’t be played",
-            "variations": [
-                "Can you explain how game access works?",
-                "I need assistance regarding game access. What should I do?",
-                "My transaction related to game access isn’t going through. Can you check?",
-                "I am facing an issue with game access. How can I resolve it?"
-            ],
-            "intent": "game_access",
-            "languages": {
-                "en": "I have a question related to game access. Can you help clarify it? | Boss, Mega can’t be played",
-                "ms": "\"mega tak boleh main bos\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Malaysia VPN\\r\\n\\r\\nhttps://play.google.com/store/apps/details?id=malaysia.vpn_tap2free&hl=en&gl=US\\r\\n\\r\\nHi, boss, try downloading and installing it, then connect to Malaysia If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"VPN malaysia \\r\\n\\r\\nhttps://play.google.com/store/apps/details?id=malaysia.vpn_tap2free&hl=en&gl=US\\r\\n\\r\\nHi bos cuba download dan pasang, lepastu sambung ke MALAYSIA\"",
-                "cn": "\"馬來西亞VPN\\r\\n\\r\\nhttps://play.google.com/store/apps/details?id=malaysia.vpn_tap2free&hl=en&gl=US\\r\\n\\r\\n嗨，老闆，試著下載並安裝，然後連接到馬來西亞\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "game_access",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820177Z",
-            "lastUpdated": "2025-02-27T06:39:18.820181Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0247",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | I want to ask, hasn’t the money been paid yet?",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | I want to ask, hasn’t the money been paid yet?",
-                "ms": "\"nak tanya ...belum bayar duit lagi ke?\"",
-                "cn": "\"想問一下，還沒付錢嗎？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please be patient, boss, we’ll check the report later If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sabar ya bos semak report nanti ya\"",
-                "cn": "\"請耐心等待，稍後會檢查報告哦，老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820211Z",
-            "lastUpdated": "2025-02-27T06:39:18.820215Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0248",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I uploaded a receipt but it’s the wrong account",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I uploaded a receipt but it’s the wrong account",
-                "ms": "\"Saya upload resit tapi salah akaun\"",
-                "cn": "\"我上傳了收據但帳戶錯了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss, this isn’t our company’s account, did you upload the wrong receipt? If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"hi boss ini bukan acc company kami ya , adakah boss silap upload resit ya ?\"",
-                "cn": "\"嗨，老闆，這不是我們公司的帳戶哦，老闆是不是上傳錯收據了？\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820242Z",
-            "lastUpdated": "2025-02-27T06:39:18.820245Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0249",
-        "question": {
-            "text": "I have a question related to credit issues. Can you help clarify it? | Hey, check my ID, where did you throw my RM 8.00 balance, you pig?",
-            "variations": [
-                "Can you explain how credit issues works?",
-                "I need assistance regarding credit issues. What should I do?",
-                "My transaction related to credit issues isn’t going through. Can you check?",
-                "I am facing an issue with credit issues. How can I resolve it?"
-            ],
-            "intent": "credit_issues",
-            "languages": {
-                "en": "I have a question related to credit issues. Can you help clarify it? | Hey, check my ID, where did you throw my RM 8.00 balance, you pig?",
-                "ms": "\"hoiii cek saya punya id d mana kamu buang saya punta balnce rm8.00 babi betul kamu\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please log in to Mega, press score log > gaming history\\r\\nScreenshot and send it here If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"minta login mega, tekan score log > gaming history\\r\\nkrinshot send sini jap\"",
-                "cn": "\"請登錄Mega，按得分記錄 > 遊戲歷史\\r\\n截圖發到這裡一下\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "credit_issues",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820253Z",
-            "lastUpdated": "2025-02-27T06:39:18.820255Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0250",
-        "question": {
-            "text": "I have a question related to credit issues. Can you help clarify it? | That doesn’t work either, you’re making me confused",
-            "variations": [
-                "Can you explain how credit issues works?",
-                "I need assistance regarding credit issues. What should I do?",
-                "My transaction related to credit issues isn’t going through. Can you check?",
-                "I am facing an issue with credit issues. How can I resolve it?"
-            ],
-            "intent": "credit_issues",
-            "languages": {
-                "en": "I have a question related to credit issues. Can you help clarify it? | That doesn’t work either, you’re making me confused",
-                "ms": "\"x boleh juga itu...kamu buat saya tidak faham\"",
-                "cn": "\"那樣也不行，你搞得我完全不明白\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Try restarting your phone, switch to another WiFi\\r\\nTry again If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"cuba restart fon , tukar wifi lain\\r\\ntry semula\"",
-                "cn": "\"試著重啟手機，換另一個WiFi\\r\\n再試一次\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "credit_issues",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820263Z",
-            "lastUpdated": "2025-02-27T06:39:18.820265Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0251",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I topped up but it hasn’t come in",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I topped up but it hasn’t come in",
-                "ms": "\"Saya topup tapi belum masuk\"",
-                "cn": "\"我充值了但還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please upload your bank slip via (WALLET -> TOPUP), the process takes 3-5 minutes, thank you, boss, awesome! ❤❤ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sila uplaod slip bankin anda melalui (WALLET -> TOPUP ) proses ambil masa dalam 3-5min ,tq boss, mantap❤❤\"",
-                "cn": "\"請通過（WALLET -> TOPUP）上傳您的銀行單據，處理時間為3-5分鐘，謝謝老闆，太棒了❤❤\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820281Z",
-            "lastUpdated": "2025-02-27T06:39:18.820283Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0252",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | I want to ask, boss, I withdrew 100, it hasn’t reached my bank yet, how many minutes do I wait, boss?",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | I want to ask, boss, I withdrew 100, it hasn’t reached my bank yet, how many minutes do I wait, boss?",
-                "ms": "\"nak tanya boss saya cuci 100 belum masok bank boss berapa minit tunggu bos.\"",
-                "cn": "\"想問，老闆，我提了100還沒進銀行，要等幾分鐘啊，老闆？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Withdrawal of winnings takes 5-30 minutes, boss, please wait a moment ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Pengeluaran wang menang ambil masa dalam 5-30 min yee boss ,sila tunggu sebentar yeee❤️\"",
-                "cn": "\"贏錢的提款需要5-30分鐘哦，老闆，請稍等一下哦❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820292Z",
-            "lastUpdated": "2025-02-27T06:39:18.820294Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0253",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, can 8 ringgit work?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, can 8 ringgit work?",
-                "ms": "\"bos 8 ringgit boleh\"",
-                "cn": "\"老闆，8令吉可以嗎？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Minimum top-up is RM 10\\r\\n\\r\\nPlease transfer another RM 10 and we’ll handle it (RM 18) (receipt claimable within 24 hours)\\r\\n\\r\\nSend both receipts here first, we’ll take care of it for you If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"minimum topup RM 10\\r\\n\\r\\nsila transfer lagi RM 10 dan kami uruskan (RM 18) ( resit claim dlm 24jam )\\r\\n\\r\\ndua2 resit hantar mari sini dulu , kami uruskan bagi ya\"",
-                "cn": "\"最低充值10令吉\\r\\n\\r\\n請再轉帳10令吉，我們會處理（18令吉）（收據在24小時內認領）\\r\\n\\r\\n先把兩張收據發到這裡，我們會幫您處理\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820307Z",
-            "lastUpdated": "2025-02-27T06:39:18.820309Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0254",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | Boss, the points for the PSV vs Juventus match haven’t come in yet",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | Boss, the points for the PSV vs Juventus match haven’t come in yet",
-                "ms": "\"bos point bola PSv vs Juventus xmasuk ya\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "All results are based on 90 minutes + injury time (excluding 15min+15min extra time and penalties after extra time) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"semua result ikut 90minutes + MASA KECEDERAAN ( TIDAK TERMASUK 15min+15min extra time dan penalty lepas extra time )\"",
-                "cn": "\"所有結果只算90分鐘 + 傷停時間（不包括15分鐘+15分鐘的加時賽和加時後的點球）\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820321Z",
-            "lastUpdated": "2025-02-27T06:39:18.820323Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0255",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | Boss, why hasn’t the winnings money entered my wallet? Jaguar 3 PM",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | Boss, why hasn’t the winnings money entered my wallet? Jaguar 3 PM",
-                "ms": "\"Boss kenapa duit kena nombor belum masuk wallet . Jaguar 3pm\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please be patient, boss, we’ll check the report later If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sabar bos semak report nanti ya\"",
-                "cn": "\"請耐心等待，老闆，稍後會檢查報告哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820332Z",
-            "lastUpdated": "2025-02-27T06:39:18.820334Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0256",
-        "question": {
-            "text": "I have a question related to game transfer. Can you help clarify it? | How do I transfer the amount?",
-            "variations": [
-                "Can you explain how game transfer works?",
-                "I need assistance regarding game transfer. What should I do?",
-                "My transaction related to game transfer isn’t going through. Can you check?",
-                "I am facing an issue with game transfer. How can I resolve it?"
-            ],
-            "intent": "game_transfer",
-            "languages": {
-                "en": "I have a question related to game transfer. Can you help clarify it? | How do I transfer the amount?",
-                "ms": "\"macam mana mahu transfer amount?\"",
-                "cn": "\"怎麼轉移金額？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Press transfer > then directly open the slot game, boss\\r\\nDON’T PRESS QUIT If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"tekan transfer > lepastu terus buka slot game bos\\r\\nJGN TEKAN QUIT\"",
-                "cn": "\"按轉移 > 然後直接打開老虎機遊戲，老闆\\r\\n不要按退出\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "game_transfer",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820361Z",
-            "lastUpdated": "2025-02-27T06:39:18.820363Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0257",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred to CIMB but selected the wrong bank",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred to CIMB but selected the wrong bank",
-                "ms": "\"Saya transfer ke CIMB tapi salah pilih bank\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "If you transferred to CIMB, please select CIMB bank and upload the receipt If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"kalau bos masuk cimb sila pilih cimb bank upload resit\"",
-                "cn": "\"如果老闆轉到CIMB，請選擇CIMB銀行上傳收據\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820397Z",
-            "lastUpdated": "2025-02-27T06:39:18.820399Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0258",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I forgot to fill in the JOKER ID during transfer",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I forgot to fill in the JOKER ID during transfer",
-                "ms": "\"Saya lupa isi ID JOKER waktu transfer\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss~\\r\\nPlease fill in your JOKER ID during online transfers in the future~\\r\\n\\r\\n❗️Fill in the designated place❗️\\r\\n✅ Recipient Reference\\r\\n✅ Other Payment Details\\r\\n\\r\\n▶️ 35559JK ◀️\\r\\n\\r\\nThank you, boss ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Hi Bosskuu ~\\r\\nMinta Masa Depan Isi ID JOKER Anda Semasa Buat Online Transfer Ya ~\\r\\n\\r\\n❗️ISI DI TEMPAT ❗️\\r\\n✅ Recipient Reference\\r\\n✅Other Payment Details\\r\\n\\r\\n▶️ 35559JK ◀️\\r\\n\\r\\nTerima Kasih Boss ❤️\"",
-                "cn": "\"嗨，老闆~\\r\\n請未來在線上轉帳時填寫您的JOKER ID哦~\\r\\n\\r\\n❗️填在指定位置❗️\\r\\n✅ 收款人參考\\r\\n✅ 其他付款詳情\\r\\n\\r\\n▶️ 35559JK ◀️\\r\\n\\r\\n謝謝老闆 ❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820411Z",
-            "lastUpdated": "2025-02-27T06:39:18.820414Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0259",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | Hi, boss, PSV vs Juventus 3-1 = 17, I bet RM 3, but boss gave it wrong, overpaid by RM 1.8",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | Hi, boss, PSV vs Juventus 3-1 = 17, I bet RM 3, but boss gave it wrong, overpaid by RM 1.8",
-                "ms": "\"hallo bos PSV vs Juventus 3-1 =17 sy bet rm 3 ini bos bagi salah sudah lebih rm 1,8\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "All results are based on 90 minutes + injury time (excluding 15min+15min extra time and penalties after extra time)\\r\\nBased on a 2-1 result If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"semua result ikut 90minutes + MASA KECEDERAAN ( TIDAK TERMASUK 15min+15min extra time dan penalty lepas extra time )\\r\\nikut result 2-1\"",
-                "cn": "\"所有結果只算90分鐘 + 傷停時間（不包括15分鐘+15分鐘加時賽和加時後的點球）\\r\\n按2-1的結果算\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820426Z",
-            "lastUpdated": "2025-02-27T06:39:18.820428Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0260",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, the amount is wrong, sorry",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, the amount is wrong, sorry",
-                "ms": "\"bos salah jumalh sori\"",
-                "cn": "\"老闆，金額錯了，抱歉\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Do you have the receipt? If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"resit ada ?\"",
-                "cn": "\"有收據嗎？\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820450Z",
-            "lastUpdated": "2025-02-27T06:39:18.820452Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0261",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I topped up the wrong amount",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I topped up the wrong amount",
-                "ms": "\"Saya topup salah jumlah\"",
-                "cn": "\"我充值金額錯了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "The amount earlier was written wrongly, please upload again and write the amount as RM 75 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Amount tadi silap tulis , sila upload semula dan amount tulis RM 75\"",
-                "cn": "\"剛才金額寫錯了，請重新上傳並將金額寫為75令吉\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820463Z",
-            "lastUpdated": "2025-02-27T06:39:18.820465Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0262",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Boss, can you check my withdrawal? It hasn’t come in yet, they said it’s approved",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Boss, can you check my withdrawal? It hasn’t come in yet, they said it’s approved",
-                "ms": "\"Bos blh check withdraw tak.. Blm masuk lagi. Sana cakap approve\"",
-                "cn": "\"老闆，能幫我檢查提款嗎？還沒進來，那邊說已批准\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user confirmed it’s in) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶確認已進入)",
-                "cn": "(無明確解決方案，僅用戶確認已進入)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820538Z",
-            "lastUpdated": "2025-02-27T06:39:18.820540Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0263",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, the receipt didn’t come up",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, the receipt didn’t come up",
-                "ms": "\"Boss resit tkda keluar\"",
-                "cn": "\"老闆，收據沒出來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, only user provided info) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，僅用戶提供資訊)",
-                "cn": "(無明確解決方案，僅用戶提供資訊)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820551Z",
-            "lastUpdated": "2025-02-27T06:39:18.820553Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0264",
-        "question": {
-            "text": "I have a question related to bank account. Can you help clarify it? | I want to add a bank account, but can I use my wife’s?",
-            "variations": [
-                "Can you explain how bank account works?",
-                "I need assistance regarding bank account. What should I do?",
-                "My transaction related to bank account isn’t going through. Can you check?",
-                "I am facing an issue with bank account. How can I resolve it?"
-            ],
-            "intent": "bank_account",
-            "languages": {
-                "en": "I have a question related to bank account. Can you help clarify it? | I want to add a bank account, but can I use my wife’s?",
-                "ms": "\"Nk tambah bank acc..tpi pakai wife punya boleh ke?\"",
-                "cn": "\"我想加銀行帳戶，但用我老婆的可以嗎？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(MUHAMMAD AMIRUL YUSOUFF) You can only use a bank under this name, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"( MUHAMMAD AMIRUL YUSOUFF ) hanya leh guna bnk atas nama ni saja ya bos\"",
-                "cn": "\"(MUHAMMAD AMIRUL YUSOUFF) 只能用這個名字的銀行哦，老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "account_management"
-            ],
-            "subCategory": "bank_account",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820567Z",
-            "lastUpdated": "2025-02-27T06:39:18.820569Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0265",
-        "question": {
-            "text": "I have a question related to game transfer. Can you help clarify it? | Boss, why is there no credit when I enter the game?",
-            "variations": [
-                "Can you explain how game transfer works?",
-                "I need assistance regarding game transfer. What should I do?",
-                "My transaction related to game transfer isn’t going through. Can you check?",
-                "I am facing an issue with game transfer. How can I resolve it?"
-            ],
-            "intent": "game_transfer",
-            "languages": {
-                "en": "I have a question related to game transfer. Can you help clarify it? | Boss, why is there no credit when I enter the game?",
-                "ms": "\"boss apasal masuk game cardit xde\"",
-                "cn": "\"老闆，為什麼進遊戲信用沒了？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Boss, after pressing transfer\\r\\nDO NOT press QUIT\\r\\nDirectly open the slot game If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"bos lepas tekan transfer\\r\\nJGN JGN tekan QUIT\\r\\nterus buka slot game\"",
-                "cn": "\"老闆，按轉移後\\r\\n千萬不要按退出\\r\\n直接打開老虎機遊戲\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "game_transfer",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820581Z",
-            "lastUpdated": "2025-02-27T06:39:18.820583Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0266",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | This one hasn’t come in yet",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | This one hasn’t come in yet",
-                "ms": "\"yg nie belum masuk\"",
-                "cn": "\"這個還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please upload your bank slip via (WALLET -> TOPUP), the process takes 3-5 minutes, thank you, boss, awesome! ❤️❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sila uplaod slip bankin anda melalui (WALLET -> TOPUP ) proses ambil masa dalam 3-5min ,tq boss, mantap❤️❤️\"",
-                "cn": "\"請通過（WALLET -> TOPUP）上傳您的銀行單據，處理時間為3-5分鐘，謝謝老闆，太棒了❤️❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820595Z",
-            "lastUpdated": "2025-02-27T06:39:18.820597Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0267",
-        "question": {
-            "text": "I have a question related to winnings. Can you help clarify it? | Why didn’t it pay?",
-            "variations": [
-                "Can you explain how winnings works?",
-                "I need assistance regarding winnings. What should I do?",
-                "My transaction related to winnings isn’t going through. Can you check?",
-                "I am facing an issue with winnings. How can I resolve it?"
-            ],
-            "intent": "winnings",
-            "languages": {
-                "en": "I have a question related to winnings. Can you help clarify it? | Why didn’t it pay?",
-                "ms": "\"kenapa tidak kena bayar\"",
-                "cn": "\"為什麼沒付錢？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Boss, please check your history\\r\\nAll records are there If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"bos semka dkt history anda ya\\r\\ndlm ada sem,ua rekod\"",
-                "cn": "\"老闆，請檢查您的歷史記錄哦\\r\\n裡面有所有記錄\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "winnings",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820604Z",
-            "lastUpdated": "2025-02-27T06:39:18.820607Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0268",
-        "question": {
-            "text": "I have a question related to game issue. Can you help clarify it? | I’ve already exited, why is my money gone, boss?",
-            "variations": [
-                "Can you explain how game issue works?",
-                "I need assistance regarding game issue. What should I do?",
-                "My transaction related to game issue isn’t going through. Can you check?",
-                "I am facing an issue with game issue. How can I resolve it?"
-            ],
-            "intent": "game_issue",
-            "languages": {
-                "en": "I have a question related to game issue. Can you help clarify it? | I’ve already exited, why is my money gone, boss?",
-                "ms": "\"sya suda kluar kenapa tiada duit saya bos\"",
-                "cn": "\"我已經退出，為什麼我的錢沒了，老闆？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Boss, please log back into the game\\r\\nJust check\\r\\nThere’s a history inside If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"bos login balik game tu\\r\\nsemak saja\\r\\ndlm ada hisotry\"",
-                "cn": "\"老闆，請重新登錄遊戲\\r\\n檢查一下\\r\\n裡面有歷史記錄\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "game_issue",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820623Z",
-            "lastUpdated": "2025-02-27T06:39:18.820626Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0269",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, if I transfer via an ATM machine and then send the receipt, is that okay?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, if I transfer via an ATM machine and then send the receipt, is that okay?",
-                "ms": "\"bos.. klau I transfer kat atm mesin.. lepas tu send resit bleh\"",
-                "cn": "\"老闆，如果我在櫃員機轉帳，然後發送收據可以嗎？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Yes, boss, just upload the receipt as usual If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"boleh bos upload resit cam biasa ja ya\"",
-                "cn": "\"可以，老闆，像平常一樣上傳收據就行哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820633Z",
-            "lastUpdated": "2025-02-27T06:39:18.820635Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0270",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Boss, there’s no money in my bank",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Boss, there’s no money in my bank",
-                "ms": "\"Boss saya puniya bank takde dueit pun\"",
-                "cn": "\"老闆，我的銀行裡沒有錢\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Withdrawal of winnings takes 5-30 minutes, boss, please wait a moment ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Pengeluaran wang menang ambil masa dalam 5-30 min yee boss ,sila tunggu sebentar yeee❤️\"",
-                "cn": "\"贏錢的提款需要5-30分鐘哦，老闆，請稍等一下哦❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820642Z",
-            "lastUpdated": "2025-02-27T06:39:18.820644Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0271",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | My money hasn’t come in yet",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | My money hasn’t come in yet",
-                "ms": "\"Blom masuk lagi dueit sya\"",
-                "cn": "\"我的錢還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please send a screenshot of your bank history here\\r\\nWe’ll get back to you within 24 hours If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"minta bos krinshot bank history mari jab\\r\\nkita akan roger boss balik dlm 24jam\"",
-                "cn": "\"請老闆截圖銀行歷史記錄發過來\\r\\n我們會在24小時內回覆老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820655Z",
-            "lastUpdated": "2025-02-27T06:39:18.820658Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0272",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Can’t you return the 450 RM to my ID now?",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Can’t you return the 450 RM to my ID now?",
-                "ms": "\"Sekrang tabole ka Kasi balik dueit 450 itu I'd sya\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "We’ll get back to you within 24 hours, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"kita akan roger boss balik dlm 24jam\"",
-                "cn": "\"我們會在24小時內回覆老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820674Z",
-            "lastUpdated": "2025-02-27T06:39:18.820676Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0273",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Boss, I withdrew 90 earlier and it hasn’t come in yet",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Boss, I withdrew 90 earlier and it hasn’t come in yet",
-                "ms": "\"Bos td sy withdraw 90 blm msk\"",
-                "cn": "\"老闆，剛才我提了90還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Withdrawal of winnings takes 5-30 minutes, boss, please wait a moment ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Pengeluaran wang menang ambil masa dalam 5-30 min yee boss ,sila tunggu sebentar yeee❤️\"",
-                "cn": "\"贏錢的提款需要5-30分鐘哦，老闆，請稍等一下哦❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820697Z",
-            "lastUpdated": "2025-02-27T06:39:18.820699Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0274",
-        "question": {
-            "text": "I have a question related to registration. Can you help clarify it? | My registered name is wrong",
-            "variations": [
-                "Can you explain how registration works?",
-                "I need assistance regarding registration. What should I do?",
-                "My transaction related to registration isn’t going through. Can you check?",
-                "I am facing an issue with registration. How can I resolve it?"
-            ],
-            "intent": "registration",
-            "languages": {
-                "en": "I have a question related to registration. Can you help clarify it? | My registered name is wrong",
-                "ms": "\"Nama pendaftaran saya salah\"",
-                "cn": "\"我的註冊姓名錯了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "\\r\\n‼️ATTENTION‼️\\r\\n➡️Your registered name is incorrect⬅️\\r\\nPlease provide your full name (FULL NAME SAME AS IC) in this chat to update your registration If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"\\r\\n‼️PERHATIAN‼️\\r\\n➡️Nama daftar anda ade silap ye ⬅️\\r\\nSila berikan nama penuh (NAMA PENUH SAMA SEPERTI IC) di chat ini untuk kemaskini nama daftar anda\"",
-                "cn": "\"\\r\\n‼️注意‼️\\r\\n➡️您的註冊姓名有誤哦⬅️\\r\\n請在此聊天中提供全名（與身份證完全相同的全名）以更新您的註冊姓名\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "account_management"
-            ],
-            "subCategory": "registration",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820717Z",
-            "lastUpdated": "2025-02-27T06:39:18.820719Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0275",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, my fast payment hasn’t come in yet?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Boss, my fast payment hasn’t come in yet?",
-                "ms": "\"bos fast payment saya blom. masok lg oo??\"",
-                "cn": "\"老闆，我的快速付款還沒進來哦？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Do you have the receipt? If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"resit ada?\"",
-                "cn": "\"有收據嗎？\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820726Z",
-            "lastUpdated": "2025-02-27T06:39:18.820728Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0276",
-        "question": {
-            "text": "I have a question related to game issue. Can you help clarify it? | Why did it fail?",
-            "variations": [
-                "Can you explain how game issue works?",
-                "I need assistance regarding game issue. What should I do?",
-                "My transaction related to game issue isn’t going through. Can you check?",
-                "I am facing an issue with game issue. How can I resolve it?"
-            ],
-            "intent": "game_issue",
-            "languages": {
-                "en": "I have a question related to game issue. Can you help clarify it? | Why did it fail?",
-                "ms": "\"why fail?\"",
-                "cn": "\"為什麼失敗？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Try transferring 0 to ID 01576023535, screenshot the record and send it here If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"cuba trasnfer 0 ke id 01576023535 krinshot screlog mari\"",
-                "cn": "\"試著轉0到ID 01576023535，截圖記錄發過來\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "game_issue",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820740Z",
-            "lastUpdated": "2025-02-27T06:39:18.820742Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0277",
-        "question": {
-            "text": "I have a question related to betting rules. Can you help clarify it? | What does 0-9 mean in 4D betting?",
-            "variations": [
-                "Can you explain how betting rules works?",
-                "I need assistance regarding betting rules. What should I do?",
-                "My transaction related to betting rules isn’t going through. Can you check?",
-                "I am facing an issue with betting rules. How can I resolve it?"
-            ],
-            "intent": "betting_rules",
-            "languages": {
-                "en": "I have a question related to betting rules. Can you help clarify it? | What does 0-9 mean in 4D betting?",
-                "ms": "\"apa maksud 0-9 dekat 4D bet?\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Bet numbers 0 to 9\\r\\nFor example, 123x means 1230 to 1239 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"bet no 0 sampai 9\\r\\ncontoh 123x maksud 1230 sampai 1239\"",
-                "cn": "\"投注號碼0到9\\r\\n例如123x，意思是1230到1239\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "game"
-            ],
-            "subCategory": "betting_rules",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820759Z",
-            "lastUpdated": "2025-02-27T06:39:18.820763Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0278",
-        "question": {
-            "text": "I have a question related to credit issues. Can you help clarify it? | The money disappeared",
-            "variations": [
-                "Can you explain how credit issues works?",
-                "I need assistance regarding credit issues. What should I do?",
-                "My transaction related to credit issues isn’t going through. Can you check?",
-                "I am facing an issue with credit issues. How can I resolve it?"
-            ],
-            "intent": "credit_issues",
-            "languages": {
-                "en": "I have a question related to credit issues. Can you help clarify it? | The money disappeared",
-                "ms": "\"hilang duit\"",
-                "cn": "\"錢不見了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Restart your phone, transfer 0 to ID 01576023535, and check again If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"restart fon trasnfer 0 ke id 01576023535 semak balik jab\"",
-                "cn": "\"重啟手機，轉0到ID 01576023535，再檢查一下\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "credit_issues",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820777Z",
-            "lastUpdated": "2025-02-27T06:39:18.820789Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0279",
-        "question": {
-            "text": "I have a question related to app issue. Can you help clarify it? | It’s still the same",
-            "variations": [
-                "Can you explain how app issue works?",
-                "I need assistance regarding app issue. What should I do?",
-                "My transaction related to app issue isn’t going through. Can you check?",
-                "I am facing an issue with app issue. How can I resolve it?"
-            ],
-            "intent": "app_issue",
-            "languages": {
-                "en": "I have a question related to app issue. Can you help clarify it? | It’s still the same",
-                "ms": "\"maseh sama\"",
-                "cn": "\"還是沒變\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Boss, try deleting the app and downloading it again If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"bos cuba padam app download semula\"",
-                "cn": "\"老闆，試著刪除 приложение 並重新下載\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "app_issue",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820822Z",
-            "lastUpdated": "2025-02-27T06:39:18.820826Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0280",
-        "question": {
-            "text": "I have a question related to network issue. Can you help clarify it? | Where’s my money? It’s still not there",
-            "variations": [
-                "Can you explain how network issue works?",
-                "I need assistance regarding network issue. What should I do?",
-                "My transaction related to network issue isn’t going through. Can you check?",
-                "I am facing an issue with network issue. How can I resolve it?"
-            ],
-            "intent": "network_issue",
-            "languages": {
-                "en": "I have a question related to network issue. Can you help clarify it? | Where’s my money? It’s still not there",
-                "ms": "\"Mana duit saya? masih tiada\"",
-                "cn": "\"我的錢呢？還是沒有\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "https://play.google.com/store/apps/details?id=malaysia.vpn_tap2free&hl=en&gl=US\\r\\n\\r\\nHi, boss, please install this Malaysia VPN and CONNECT TO MALAYSIA\\r\\n\\r\\nInstall this and try again If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"https://play.google.com/store/apps/details?id=malaysia.vpn_tap2free&hl=en&gl=US\\r\\n\\r\\nHi Bos, sila pasang VPN Malaysia ini dan CONNECT TO MALAYSIA\\r\\n\\r\\npasang ni dan cuba semula\"",
-                "cn": "\"https://play.google.com/store/apps/details?id=malaysia.vpn_tap2free&hl=en&gl=US\\r\\n\\r\\n嗨，老闆，請安裝這個馬來西亞VPN並連接到馬來西亞\\r\\n\\r\\n安裝這個再試一次\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "network_issue",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820848Z",
-            "lastUpdated": "2025-02-27T06:39:18.820853Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0281",
-        "question": {
-            "text": "I have a question related to credit issues. Can you help clarify it? | Why did 2000 disappear?",
-            "variations": [
-                "Can you explain how credit issues works?",
-                "I need assistance regarding credit issues. What should I do?",
-                "My transaction related to credit issues isn’t going through. Can you check?",
-                "I am facing an issue with credit issues. How can I resolve it?"
-            ],
-            "intent": "credit_issues",
-            "languages": {
-                "en": "I have a question related to credit issues. Can you help clarify it? | Why did 2000 disappear?",
-                "ms": "\"kenapa hilang 2000?\"",
-                "cn": "\"為什麼2000不見了？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Boss, screenshot the Joker history and send it here If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"bos krinshot history joker mari jab\"",
-                "cn": "\"老闆，截圖Joker的歷史記錄發過來\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "credit_issues",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820871Z",
-            "lastUpdated": "2025-02-27T06:39:18.820875Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0282",
-        "question": {
-            "text": "I have a question related to app issue. Can you help clarify it? | What’s going on? Where’s the money?",
-            "variations": [
-                "Can you explain how app issue works?",
-                "I need assistance regarding app issue. What should I do?",
-                "My transaction related to app issue isn’t going through. Can you check?",
-                "I am facing an issue with app issue. How can I resolve it?"
-            ],
-            "intent": "app_issue",
-            "languages": {
-                "en": "I have a question related to app issue. Can you help clarify it? | What’s going on? Where’s the money?",
-                "ms": "\"MACAMNA? MANA DUIT?\"",
-                "cn": "\"怎麼回事？錢呢？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Boss, try logging out of the app and logging back in If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"bos cuba logout app dan login balik\"",
-                "cn": "\"老闆，試著登出應用程式然後重新登入\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "app_issue",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820888Z",
-            "lastUpdated": "2025-02-27T06:39:18.820892Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0283",
-        "question": {
-            "text": "I have a question related to credit issues. Can you help clarify it? | Boss, don’t lie to me",
-            "variations": [
-                "Can you explain how credit issues works?",
-                "I need assistance regarding credit issues. What should I do?",
-                "My transaction related to credit issues isn’t going through. Can you check?",
-                "I am facing an issue with credit issues. How can I resolve it?"
-            ],
-            "intent": "credit_issues",
-            "languages": {
-                "en": "I have a question related to credit issues. Can you help clarify it? | Boss, don’t lie to me",
-                "ms": "\"bos jangan tipu\"",
-                "cn": "\"老闆別騙我\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "According to the scorelog, boss, you’ve used it all up If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"ikut scorelog bos da habis guna ya\"",
-                "cn": "\"根據記錄，老闆已經用完了哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "credit_issues",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820899Z",
-            "lastUpdated": "2025-02-27T06:39:18.820901Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0284",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | How does it come in?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | How does it come in?",
-                "ms": "\"Macam mana masuk\"",
-                "cn": "\"怎麼進來的？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Press the Maybank box\\r\\nand upload the receipt If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"tekan kotak maybank\\r\\ndan uploadkan resit\"",
-                "cn": "\"按Maybank的框框\\r\\n然後上傳收據\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820917Z",
-            "lastUpdated": "2025-02-27T06:39:18.820920Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0285",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | The credit hasn’t come in yet",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | The credit hasn’t come in yet",
-                "ms": "\"kredit belum masuk\"",
-                "cn": "\"信用還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Send the receipt here If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"resit send sini\"",
-                "cn": "\"把收據發到這裡\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820929Z",
-            "lastUpdated": "2025-02-27T06:39:18.820931Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0286",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | There’s a CIMB top-up of RM 200",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | There’s a CIMB top-up of RM 200",
-                "ms": "\"ada top-up cimb RM 200\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please upload your bank slip via (WALLET -> TOPUP), the process takes 3-5 minutes, thank you, boss, awesome! ❤️❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sila uplaod slip bankin anda melalui (WALLET -> TOPUP ) proses ambil masa dalam 3-5min ,tq boss, mantap❤️❤️\"",
-                "cn": "\"請通過（WALLET -> TOPUP）上傳您的銀行單據，處理時間為3-5分鐘，謝謝老闆，太棒了❤️❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820942Z",
-            "lastUpdated": "2025-02-27T06:39:18.820944Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0287",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Top-up is done",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Top-up is done",
-                "ms": "\"done topup\"",
-                "cn": "\"充值完成了\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Please upload your bank slip via (WALLET -> TOPUP), the process takes 3-5 minutes, thank you, boss, awesome! ❤️❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"sila uplaod slip bankin anda melalui (WALLET -> TOPUP ) proses ambil masa dalam 3-5min ,tq boss, mantap❤️❤️\"",
-                "cn": "\"請通過（WALLET -> TOPUP）上傳您的銀行單據，處理時間為3-5分鐘，謝謝老闆，太棒了❤️❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820955Z",
-            "lastUpdated": "2025-02-27T06:39:18.820958Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0288",
-        "question": {
-            "text": "I have a question related to schedule inquiry. Can you help clarify it? | Is there a special race?",
-            "variations": [
-                "Can you explain how schedule inquiry works?",
-                "I need assistance regarding schedule inquiry. What should I do?",
-                "My transaction related to schedule inquiry isn’t going through. Can you check?",
-                "I am facing an issue with schedule inquiry. How can I resolve it?"
-            ],
-            "intent": "schedule_inquiry",
-            "languages": {
-                "en": "I have a question related to schedule inquiry. Can you help clarify it? | Is there a special race?",
-                "ms": "\"ada special race ka\"",
-                "cn": "\"有特別比賽嗎？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Today there’s GD Lotto, 9D Lotto, J2J3, Perdana, and Lucky If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Hari ini GD Lotto, 9D Lotto, J2J3, Perdana dan Lucky\"",
-                "cn": "\"今天有GD Lotto、9D Lotto、J2J3、Perdana和Lucky\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "game"
-            ],
-            "subCategory": "schedule_inquiry",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820967Z",
-            "lastUpdated": "2025-02-27T06:39:18.820969Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0289",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I forgot to include the ID during top-up",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I forgot to include the ID during top-up",
-                "ms": "\"Saya lupa letak ID waktu topup\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Hi, boss~\\r\\nPlease fill in your JOKER ID during online transfers in the future~\\r\\n\\r\\n❗️Fill in the designated place❗️\\r\\n✅ Recipient Reference\\r\\n✅ Other Payment Details\\r\\n\\r\\n▶️ 58411JK ◀️\\r\\n\\r\\nThank you, boss ❤️ If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Hi Bosskuu ~\\r\\nMinta Masa Depan Isi ID JOKER Anda Semasa Buat Online Transfer Ya ~\\r\\n\\r\\n❗️ISI DI TEMPAT ❗️\\r\\n✅ Recipient Reference\\r\\n✅Other Payment Details\\r\\n\\r\\n▶️ 58411JK ◀️\\r\\n\\r\\nTerima Kasih Boss ❤️\"",
-                "cn": "\"嗨，老闆~\\r\\n請未來在線上轉帳時填寫您的JOKER ID哦~\\r\\n\\r\\n❗️填在指定位置❗️\\r\\n✅ 收款人參考\\r\\n✅ 其他付款詳情\\r\\n\\r\\n▶️ 58411JK ◀️\\r\\n\\r\\n謝謝老闆 ❤️\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820977Z",
-            "lastUpdated": "2025-02-27T06:39:18.820980Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0290",
-        "question": {
-            "text": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Boss, the withdrawal hasn’t come in",
-            "variations": [
-                "My withdrawal request is delayed. Can you provide an update?",
-                "I tried withdrawing funds, but the money hasn’t arrived yet. What should I do?",
-                "Why is my withdrawal still pending?",
-                "I withdrew using withdraw, but the transaction is stuck. How can I fix this?"
-            ],
-            "intent": "withdraw",
-            "languages": {
-                "en": "I requested a withdrawal, but the funds have not been received yet. What could be the issue? | Boss, the withdrawal hasn’t come in",
-                "ms": "\"bos wd x masuk\"",
-                "cn": "\"老闆，提款沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Send the receipt here If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"resit send sini\"",
-                "cn": "\"把收據發到這裡\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "withdraw",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.820998Z",
-            "lastUpdated": "2025-02-27T06:39:18.821002Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0291",
-        "question": {
-            "text": "I have a question related to password. Can you help clarify it? | What’s the password, boss?",
-            "variations": [
-                "Can you explain how password works?",
-                "I need assistance regarding password. What should I do?",
-                "My transaction related to password isn’t going through. Can you check?",
-                "I am facing an issue with password. How can I resolve it?"
-            ],
-            "intent": "password",
-            "languages": {
-                "en": "I have a question related to password. Can you help clarify it? | What’s the password, boss?",
-                "ms": "\"password apa ya bos\"",
-                "cn": "\"密碼是什麼啊，老闆？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Boss, copy it from the app If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"bos copy dlm app ya\"",
-                "cn": "\"老闆，在應用程式裡複製哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "password",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821021Z",
-            "lastUpdated": "2025-02-27T06:39:18.821024Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0292",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Top-up via Touch n Go hasn’t come in yet",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Top-up via Touch n Go hasn’t come in yet",
-                "ms": "\"topup via touch n go x msuk lg\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Wait 5 minutes, if it doesn’t come in, contact us again, boss If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"tunggu 5min tak masuk roger balik ye bos\"",
-                "cn": "\"等5分鐘如果沒進來請再聯繫哦，老闆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821035Z",
-            "lastUpdated": "2025-02-27T06:39:18.821037Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0293",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | It hasn’t come in yet, boss",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | It hasn’t come in yet, boss",
-                "ms": "\"Belum masuk lagi bossku\"",
-                "cn": "\"還沒進來，老闆\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "How to Retrieve Transfer Receipt from MyBSN\\r\\nhttps://www.youtube.com/watch?v=gIVHfyD_pIE\\r\\nFollow this method to screenshot the full receipt If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Cara Dapatkan Semula Resit Transfer Dari MyBSN\\r\\nhttps://www.youtube.com/watch?v=gIVHfyD_pIE\\r\\nikut cara ni krinshot full resit\"",
-                "cn": "\"如何從MyBSN重新獲取轉帳收據\\r\\nhttps://www.youtube.com/watch?v=gIVHfyD_pIE\\r\\n按照這個方法截圖完整收據\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821053Z",
-            "lastUpdated": "2025-02-27T06:39:18.821056Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0294",
-        "question": {
-            "text": "I have a question related to app issue. Can you help clarify it? | Is Joker under maintenance?",
-            "variations": [
-                "Can you explain how app issue works?",
-                "I need assistance regarding app issue. What should I do?",
-                "My transaction related to app issue isn’t going through. Can you check?",
-                "I am facing an issue with app issue. How can I resolve it?"
-            ],
-            "intent": "app_issue",
-            "languages": {
-                "en": "I have a question related to app issue. Can you help clarify it? | Is Joker under maintenance?",
-                "ms": "\"joker maintenance ka\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Restart your phone and check again If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"restart fon semak balik\"",
-                "cn": "\"重啟手機再檢查一下\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "app_issue",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821074Z",
-            "lastUpdated": "2025-02-27T06:39:18.821078Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0295",
-        "question": {
-            "text": "I have a question related to credit issues. Can you help clarify it? | I’ve deleted Joker and reinstalled it, but the credit is still 0",
-            "variations": [
-                "Can you explain how credit issues works?",
-                "I need assistance regarding credit issues. What should I do?",
-                "My transaction related to credit issues isn’t going through. Can you check?",
-                "I am facing an issue with credit issues. How can I resolve it?"
-            ],
-            "intent": "credit_issues",
-            "languages": {
-                "en": "I have a question related to credit issues. Can you help clarify it? | I’ve deleted Joker and reinstalled it, but the credit is still 0",
-                "ms": "\"dah padam joker install balik tapi kredit masih 0\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "https://4djokers.club/chat/download/joker\\r\\nBoss, try deleting and downloading it again If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"https://4djokers.club/chat/download/joker\\r\\nbos cuba padam download semula\"",
-                "cn": "\"https://4djokers.club/chat/download/joker\\r\\n老闆試著刪除並重新下載\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "credit_issues",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821096Z",
-            "lastUpdated": "2025-02-27T06:39:18.821100Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0296",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Why hasn’t the money come in yet?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | Why hasn’t the money come in yet?",
-                "ms": "\"Knpa duit blum masuk lagi\"",
-                "cn": "\"為什麼錢還沒進來？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, user later confirmed it’s in) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，用戶後續確認已進入)",
-                "cn": "(無明確解決方案，用戶後續確認已進入)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821137Z",
-            "lastUpdated": "2025-02-27T06:39:18.821140Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0297",
-        "question": {
-            "text": "I have a question related to credit transfer error. Can you help clarify it? | I mistakenly shared credit with someone else’s ID earlier",
-            "variations": [
-                "Can you explain how credit transfer error works?",
-                "I need assistance regarding credit transfer error. What should I do?",
-                "My transaction related to credit transfer error isn’t going through. Can you check?",
-                "I am facing an issue with credit transfer error. How can I resolve it?"
-            ],
-            "intent": "credit_transfer_error",
-            "languages": {
-                "en": "I have a question related to credit transfer error. Can you help clarify it? | I mistakenly shared credit with someone else’s ID earlier",
-                "ms": "\"sy salah credit share sma id org lain td\"",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Send the screenshot here If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"krinshot send sini\"",
-                "cn": "\"把截圖發到這裡\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "credit_transfer_error",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821154Z",
-            "lastUpdated": "2025-02-27T06:39:18.821157Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0298",
-        "question": {
-            "text": "I have a question related to credit transfer error. Can you help clarify it? | Can it be refunded?",
-            "variations": [
-                "Can you explain how credit transfer error works?",
-                "I need assistance regarding credit transfer error. What should I do?",
-                "My transaction related to credit transfer error isn’t going through. Can you check?",
-                "I am facing an issue with credit transfer error. How can I resolve it?"
-            ],
-            "intent": "credit_transfer_error",
-            "languages": {
-                "en": "I have a question related to credit transfer error. Can you help clarify it? | Can it be refunded?",
-                "ms": "\"boleh refund ka?\"",
-                "cn": "\"可以退款嗎？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "We’ve PMed him to check and return it to you, boss. Meanwhile, please wait for him to read and act on it If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"kita dah pm minta dia semak dan pulangkan balik kpd boss. sementara tu minta boss tunggu dia baca dan buat dulu ya\"",
-                "cn": "\"我們已經私訊他，請他檢查並退還給老闆。在此期間，請老闆先等他查看並處理哦\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "technical"
-            ],
-            "subCategory": "credit_transfer_error",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821176Z",
-            "lastUpdated": "2025-02-27T06:39:18.821180Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0299",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred yesterday, but it hasn’t come in until now",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | I transferred yesterday, but it hasn’t come in until now",
-                "ms": "\"sda aku transfer hri blm msk smpai skrng\"",
-                "cn": "\"我昨天已經轉帳了，到現在還沒進來\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "Now please screenshot our conversation\\r\\n\\r\\nand upload it to the top-up section as a receipt, and enter the amount RM 20\\r\\n\\r\\n\\r\\nscreenshot 👆👆👆 If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "\"Skrg minta boss screenshot perbualan kami\\r\\n\\r\\ndan upload dekat bahagian topup sebagai resit , dan isi nilai RM 20\\r\\n\\r\\n\\r\\nscreenshot 👆👆👆\"",
-                "cn": "\"現在請老闆截圖我們的對話\\r\\n\\r\\n並上傳到充值部分作為收據，填入金額20令吉\\r\\n\\r\\n\\r\\n截圖 👆👆👆\""
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.821193Z",
-            "lastUpdated": "2025-02-27T06:39:18.821196Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-    {
-        "id": "0300",
-        "question": {
-            "text": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | It hasn’t been paid yet, boss?",
-            "variations": [
-                "My topup transaction is delayed. Can you check it?",
-                "I uploaded my receipt, but my topup balance is still missing. Can you assist?",
-                "Why hasn’t my topup request been processed yet?",
-                "I made a topup payment, but it hasn’t reflected. What steps should I take?"
-            ],
-            "intent": "topup",
-            "languages": {
-                "en": "I have made a topup transaction, but the balance has not been credited yet. What should I do? | It hasn’t been paid yet, boss?",
-                "ms": "\"belum di bayar ya bos\"",
-                "cn": "\"還沒支付嗎，老闆？\""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": "(No explicit solution, no further confirmation from user) If the issue persists, please provide more details or contact support for further assistance.",
-                "ms": "(無明確解決方案，用戶未提供後續確認)",
-                "cn": "(無明確解決方案，用戶未提供後續確認)"
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": [
-                "finance"
-            ],
-            "subCategory": "topup",
-            "difficulty": 2,
-            "confidence": 0.5,
-            "dateCreated": "2025-02-27T06:39:18.822843Z",
-            "lastUpdated": "2025-02-27T06:39:18.822853Z",
-            "version": "1.0",
-            "source": "customer_service",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": null
-        }
-    },
-]
