+++ api/views/conversation_view.py
@@ -4,13 +4,9 @@ from rest_framework import status
 
 import logging
 from datetime import datetime
-import time
-
-from rapidfuzz import fuzz
 
 from api.chatbot import (
     check_answer_mongo_and_openai,
-    extrair_knowledge_items,
     get_store,
 )
 
@@ -24,7 +20,7 @@ from ..serializers import (
     PromptConversationAdminSerializer,
 )
 
-from ai_config.ai_constants import (
+from api.constants.ai_constants import (
     LANGUAGE_DEFAULT,
 )
 
@@ -134,13 +130,6 @@ class PromptConversationAdminView(APIView):
         logger.info("Starting prompt_conversation_admin request")
 
         try:
-            # Get the header value as a string
-            use_mongo_str = request.GET.get(
-                "use_mongo", "0"
-            )  # Default to "0" if not provided
-            use_mongo = use_mongo_str in ("1", "true", "yes")
-            print("Mongo: " + str(use_mongo))
-
             # Language
             language = request.GET.get("language", LANGUAGE_DEFAULT)
             print("Language " + language)
@@ -159,40 +148,11 @@ class PromptConversationAdminView(APIView):
                 )
 
             # Extract validated data
-            prompt = input_serializer.validated_data["prompt"]
-            conversation_id = input_serializer.validated_data["conversation_id"]
-            # user_id = input_serializer.validated_data["user_id"]
-
-            if use_mongo:
-                print("Using Mongo DB")
-                # Search for the answer on mongo db
-                response = fuzzy_match_with_dynamic_context(
-                    self=self,
-                    query=prompt,
-                    collection_name="feedback_data_" + language,
-                    threshold=10,
-                    language=language,
-                )
-                if response:
-                    print("Correct answer found in Mongo DB")
-                    response_data = {
-                        "generation": response,
-                        "conversation_id": conversation_id,
-                        "is_last_message": "false",
-                        "language": language,
-                    }
-                    # time.sleep(6)
-                    return Response(response_data, status=status.HTTP_200_OK)
-
-            # Extract validated data
             validated_data = input_serializer.validated_data
 
-            # Generate AI response
-            # generation_start = time.time()
             logger.info("Starting AI response generation")
 
             response = prompt_conversation_admin(
-                self,
                 user_prompt=validated_data["prompt"],
                 conversation_id=validated_data["conversation_id"],
                 admin_id=validated_data.get("admin_id", ""),
@@ -257,13 +217,6 @@ class PromptConversationView(APIView):
         logger.info("Starting prompt_conversation request")
 
         try:
-            # Get the header value as a string
-            use_mongo_str = request.GET.get(
-                "use_mongo", "0"
-            )  # Default to "0" if not provided
-            use_mongo = use_mongo_str in ("1", "true", "yes")
-            print("Mongo: " + str(use_mongo))
-
             # Language
             language = request.GET.get("language", LANGUAGE_DEFAULT)
             print("Language " + language)
@@ -279,33 +232,12 @@ class PromptConversationView(APIView):
                 return Response(
                     {"error": "Invalid input data", "details": input_serializer.errors},
                     status=status.HTTP_400_BAD_REQUEST,
-                )
-
-            # Extract validated data
-            prompt = input_serializer.validated_data["prompt"]
-
-            if use_mongo:
-                print("Using Mongo DB")
-                # Search for the answer on mongo db
-                response = fuzzy_match_with_dynamic_context(
-                    self=self,
-                    query=prompt,
-                    collection_name="feedback_data_" + language,
-                    threshold=10,
-                    language=language,
-                )
-                if response:
-                    response_data = {
-                        "generation": response,
-                    }
-                    print("Correct answer found in Mongo DB")
-                    return Response(response_data, status=status.HTTP_200_OK)
+                )         
 
             # Extract validated data
             validated_data = input_serializer.validated_data
 
             # Generate AI response
-            # generation_start = time.time()
             logger.info("Starting AI response generation")
 
             response = prompt_conversation(
@@ -364,322 +296,3 @@ class PromptConversationView(APIView):
                 {"error": f"Failed to retrieve conversation history: {str(e)}"},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
-        
-
-class ConversationDetailView(APIView):
-    def get(self, request, conversation_id):
-        db = None
-        start_time = time.time()
-        try:
-            db = MongoDB.get_db()
-            
-            conversation = db.conversations.find_one({"session_id": conversation_id})            
-            if not conversation:
-                return Response(
-                    {"error": "Conversation not found"},
-                    status=status.HTTP_404_NOT_FOUND
-                )
-            
-            # Converte o ObjectId para string
-            conversation["_id"] = str(conversation["_id"])
-            
-            return Response(conversation, status=status.HTTP_200_OK)
-        except Exception as e:
-            return Response(
-                {"error": f"Error retrieving conversation: {str(e)}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-        
-
-class AllConversationsIdsView(APIView):
-    def get(self, request):
-        db = None
-        try:
-            db = MongoDB.get_db()
-            sessions = list(db.conversations.find({}, {"session_id": 1, "_id": 0}))
-            return Response(sessions, status=status.HTTP_200_OK)
-        except Exception as e:
-            return Response(
-                {"error": f"Error retrieving session ids: {str(e)}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-        
-
-
-class UpdateKnowledgeView(APIView):
-    def post(self, request):
-        db = None
-        try:
-            conversation_id = request.data.get("conversation_id")
-            if not conversation_id:
-                return Response(
-                    {"error": "conversation_id is required"},
-                    status=status.HTTP_400_BAD_REQUEST
-                )
-            
-            db = MongoDB.get_db()
-            # Busca a conversa usando o session_id
-            conversation = db.conversations.find_one({"session_id": conversation_id})
-            if not conversation:
-                return Response(
-                    {"error": "Conversation not found"},
-                    status=status.HTTP_404_NOT_FOUND
-                )
-            
-            knowledge_items = extrair_knowledge_items(conversation)
-            
-            
-            # Retorna os itens extraídos (candidatos) sem verificar no brain
-            return Response({
-                "message": "Candidate knowledge items extracted.",
-                "candidate_items": knowledge_items,
-                "count": len(knowledge_items)
-            }, status=status.HTTP_200_OK)
-            
-        except Exception as e:
-            return Response(
-                {"error": f"Error extracting candidate knowledge: {str(e)}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-        
-
-class DeleteConversationView(APIView):
-    def delete(self, request, *args, **kwargs):
-        # Obtém o conversation_id dos kwargs (já que está na URL)
-        conversation_id = kwargs.get("conversation_id")
-        if not conversation_id:
-            return Response(
-                {"error": "conversation_id is required"},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        db = None
-        try:
-            db = MongoDB.get_db()
-            result = db.conversations.delete_one({"session_id": conversation_id})
-            if result.deleted_count == 0:
-                return Response(
-                    {"error": "Conversation not found"},
-                    status=status.HTTP_404_NOT_FOUND
-                )
-            return Response(
-                {"message": "Conversation deleted successfully"},
-                status=status.HTTP_200_OK
-            )
-        except Exception as e:
-            return Response(
-                {"error": f"Error deleting conversation: {str(e)}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-        
-
-class FinaliseConversationView(APIView):
-    def post(self, request, *args, **kwargs):
-        # Get the conversation_id from the URL kwargs
-        conversation_id = kwargs.get("conversation_id")
-        if not conversation_id:
-            return Response(
-                {"error": "conversation_id is required"},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        db = None
-        try:
-            db = MongoDB.get_db()
-            # Update the conversation document by setting 'status' to 'done'
-            result = db.conversations.update_one(
-                {"session_id": conversation_id},
-                {"$set": {"status": "done"}}
-            )
-            if result.matched_count == 0:
-                return Response(
-                    {"error": "Conversation not found"},
-                    status=status.HTTP_404_NOT_FOUND
-                )
-            return Response(
-                {"message": "Conversation finalized successfully"},
-                status=status.HTTP_200_OK
-            )
-        except Exception as e:
-            return Response(
-                {"error": f"Error finalizing conversation: {str(e)}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-        
-                
-class FinaliseAllConversationsView(APIView):
-    """
-    This view finalizes all conversations by setting their status to "done".
-    """
-    def post(self, request, *args, **kwargs):
-        db = None
-        try:
-            db = MongoDB.get_db()
-            # Update all documents in the "conversations" collection,
-            # setting the "status" field to "done"
-            result = db.conversations.update_many(
-                {},
-                {"$set": {"status": "done"}}
-            )
-            return Response(
-                {
-                    "message": "All conversations finalized successfully.",
-                    "modified_count": result.modified_count
-                },
-                status=status.HTTP_200_OK
-            )
-        except Exception as e:
-            return Response(
-                {"error": f"Error finalizing all conversations: {str(e)}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-        
-
-def categorize_conversation_resolution(conversation: dict, db) -> dict:
-    """
-    Analyzes the conversation by extracting knowledge items.
-    
-    If candidate items are present:
-      - If the short answer indicates that there is no relevant resolution,
-        the conversation is categorized as "useless": it is inserted into the
-        'read_conversation_useless' collection and removed from 'conversations'.
-      - Else, if the candidate item has a non-empty 'raw' field,
-        the candidate answer is added to the conversation, which is then moved
-        to 'read_conversation_into_brain' and deleted from 'conversations'.
-      - Else if the 'raw' field is empty,
-        the conversation is updated with the candidate answer, moved to the
-        'read_conversation_no_brain' collection, and deleted from 'conversations'.
-        
-    Returns:
-        A dictionary representing the candidate answer if a relevant resolution is found,
-        or an empty dictionary if the conversation is categorized as useless.
-    """
-    candidate_items = extrair_knowledge_items(conversation)
-    
-    if candidate_items:
-        candidate_item = candidate_items[0]
-        answer = candidate_item.get("answer", {})
-        # Retrieve and standardize the short answer text
-        short_text = answer.get("detailed", {}).get("en", "").strip().lower()
-        
-        # If the short answer explicitly indicates no relevant resolution
-        if short_text == "there is no relevant resolution in this conversation.":
-            db.read_conversation_useless.insert_one(conversation)
-            db.conversations.delete_one({"session_id": conversation["session_id"]})
-            return {}
-        
-        # Add the candidate answer into the conversation object.
-        conversation["extracted_answer"] = answer
-        
-        # Check the 'raw' field in the candidate answer
-        raw_value = answer.get("raw")
-        if raw_value:
-            # If 'raw' is not empty, move to 'read_conversation_into_brain'
-            db.read_conversation_into_brain.insert_one(conversation)
-        else:
-            # If 'raw' is empty, move to 'read_conversation_no_brain'
-            db.read_conversation_no_brain.insert_one(conversation)
-            
-        # Remove the conversation from the main 'conversations' collection
-        db.conversations.delete_one({"session_id": conversation["session_id"]})
-        
-        return answer
-    else:
-        # If no candidate items were extracted, treat the conversation as useless.
-        db.read_conversation_useless.insert_one(conversation)
-        db.conversations.delete_one({"session_id": conversation["session_id"]})
-        return {}
-
-class SeparateConversationsView(APIView):
-    """
-    This view retrieves all conversation session IDs from the "conversations" collection.
-    For each conversation with status "done", it processes the conversation using the
-    categorize_conversation_resolution function and moves it to the appropriate collection.
-    It returns an array that shows which session ID was moved to which collection.
-    """
-    def get(self, request):
-        db = None
-        results = []
-        try:
-            db = MongoDB.get_db()
-            # Retrieve only the "session_id" field from all conversations
-            sessions = list(db.conversations.find({}, {"session_id": 1, "_id": 0}))
-            logger.debug(f"Retrieved session IDs: {sessions}")
-
-            for session in sessions:
-                session_id = session.get("session_id")
-                if not session_id:
-                    logger.warning("Skipping a conversation with missing session_id.")
-                    continue
-
-                # Retrieve the full conversation based on session_id
-                conversation = db.conversations.find_one({"session_id": session_id})
-                if conversation is None:
-                    logger.warning(f"Conversation with session_id {session_id} not found.")
-                    continue
-
-                # Process only conversations with status "done"
-                if conversation.get("status") != "done":
-                    logger.debug(f"Skipping conversation {session_id} with status: {conversation.get('status')}")
-                    continue
-
-                # Call the categorization method to process the conversation.
-                candidate_answer = categorize_conversation_resolution(conversation, db)
-
-                # Determine the category based on the candidate answer:
-                if candidate_answer == {}:
-                    category = "useless"
-                else:
-                    if candidate_answer.get("raw"):
-                        category = "into_brain"
-                    else:
-                        category = "no_brain"
-
-                results.append({
-                    "session_id": session_id,
-                    "category": category
-                })
-                logger.debug(f"Processed conversation {session_id} categorized as {category}")
-
-            return Response({
-                "message": "Conversations processed successfully.",
-                "results": results
-            }, status=status.HTTP_200_OK)
-        except Exception as e:
-            logger.exception("Error processing conversations")
-            return Response({
-                "error": f"Error processing conversations: {str(e)}"
-            }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        
-
-class DashboardCountsView(APIView):
-    """
-    This endpoint returns the counts for:
-      - Total conversations
-      - Useless conversations
-      - No Brain conversations
-      - Into Brain conversations
-      - Total knowledge in brain (sum of the three latter)
-    """
-    def get(self, request):
-        db = None
-        try:
-            db = MongoDB.get_db()
-            conversations_count = db.conversations.count_documents({})
-            useless_count = db.read_conversation_useless.count_documents({})
-            no_brain_count = db.read_conversation_no_brain.count_documents({})
-            into_brain_count = db.read_conversation_into_brain.count_documents({})
-            brain_count =  get_document_count()
-
-            data = {
-                "conversations": conversations_count,
-                "useless": useless_count,
-                "noBrain": no_brain_count,
-                "intoBrain": into_brain_count,
-                "brainCount": brain_count,
-            }
-            return Response(data, status=status.HTTP_200_OK)
-        except Exception as e:
-            return Response(
-                {"error": f"Error retrieving dashboard counts: {str(e)}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-
