+++ chatbot_project/settings.py
@@ -64,7 +64,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "rest_framework",
-    "api.apps.ApiConfig",
+    "api.config.apps.ApiConfig",
     "corsheaders",
 ]
